requirements.txt
alembic==1.6.5
aniso8601==9.0.1
click==8.0.1
colorama==0.4.4
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.2.1
Flask-RESTful==0.3.9
greenlet==1.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
PyJWT==2.1.0
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.18
Werkzeug==2.0.1
